requirements.txt
streamlit>=1.45
openai>=1.0
youtube-transcript-api>=1.2
python-dotenv>=1.0
textstat>=0.7